Save and restore the noise profile with the plugin state
Implement the State LV2 extension so that the noise print can be saved
along with the session and reloaded when opening the session. It removes
the need to automate the capture button on an empty part, or any similar
workaround.

This also needs the URID LV2 extension to use standard atom types and
register state elements.