Apply makeup on the processed signal, not the dry one
Currently, the makeup gain is applied on the dry signal, which is
supposed to be passed as-is when the plugin is bypassed. That doesn't
make it transparent when (soft) bypass is enabled, and in the normal usage case where the full wet signal is used (99.9% of the time), the makeup gain control is ineffective because the wet signal is not amplified.

Apply the makeup gain on the wet signal as intended.